requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.9.2        
langchain==0.3.27
langchain-openai==0.3.32
python-dotenv==1.0.0
urllib3<2.0             